+++ .gitignore
@@ -39,3 +39,5 @@ screenshot
 .eslintcache
 
 build
+
+/src/services/supabase/key.ts
