+++ src/services/supabase/key.demo.ts
@@ -1,0 +1,3 @@
+export const supabaseUrl = 'yourSupabaseUrl';
+
+export const supabaseAnonKey = 'yourSupabaseAnonKey';
